Support round-trip updates
Converted to use ruamel.yaml so that after updates to the input YAML
the format input YAML is preserved in the final YAML. That includes
the order of the tags and comments.

That conversion expanded the domain of YAML input formats accepted to
include both strings and Path()-based file names.

Also updated a variety of the interface methods.

Finally, enchanced the examples.